# contracts/openapi/console-api.yaml
openapi: 3.0.3
info:
  title: Dynamiq Console API
  description: >
    Control-plane API served by console-api (Phalcon). PRD-001 §9.1.
    Every /admin/v1/* route except /me and /companies requires the X-Company-Id header.
  version: "0.1.0"
servers:
  - url: /admin/v1
security:
  - ssoJwt: []
components:
  securitySchemes:
    ssoJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: SSO JWT issued by the existing Phalcon system (RS256 + JWKS), PRD §7.1/7.6.
  parameters:
    XCompanyId:
      name: X-Company-Id
      in: header
      required: true
      schema:
        type: string
      description: Company context for this request; checked against membership (PRD §7.2).
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    SkillVersion:
      type: object
      properties:
        id: { type: string }
        version_no: { type: integer }
        content: { type: string, description: 'SKILL.md content, max 100 KB (PRD §6.6a).' }
        content_hash: { type: string }
        has_scripts: { type: boolean }
        is_published: { type: boolean }
        published_at: { type: string, nullable: true }
    Skill:
      type: object
      description: >-
        CRUD + versioning only (PRD §6.6a) — not yet wired into the agent spec/compiler, so
        no agent can load one via SkillsTool during a run.
      properties:
        id: { type: string }
        company_id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        status: { type: string }
        latest_version: { $ref: '#/components/schemas/SkillVersion' }
    SkillInput:
      type: object
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        status: { type: string }
        content: { type: string, description: 'Creating a new draft, or editing an already-published one, bumps the version.' }
...
    AgentVersion:
      type: object
      properties:
        id: { type: string }
        company_id: { type: string }
        agent_id: { type: string }
        version_no: { type: integer }
        spec: { type: object, additionalProperties: true }
        spec_version: { type: string }
        compiled_definition: { type: object, additionalProperties: true, nullable: true }
        compiler_version: { type: string, nullable: true }
        dynamiq_version: { type: string, nullable: true }
        is_published: { type: boolean }
        published_by: { type: string, nullable: true }
        published_at: { type: string, format: date-time, nullable: true }
...
paths:
...
  /agents/{id}/versions/{vid}/publish:
    parameters:
      - { $ref: '#/components/parameters/XCompanyId' }
      - { $ref: '#/components/parameters/IdParam' }
      - { name: vid, in: path, required: true, schema: { type: string } }
    post:
      operationId: publishAgentVersion
      summary: Recompiles via ai-runtime, sets is_published=true, bumps dependent deployments' config_version (PRD §4.4-A).
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AgentVersion' } } } }
...
  /skills:
    get:
      operationId: listSkills
      parameters: [{ $ref: '#/components/parameters/XCompanyId' }]
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Skill' } } } } } }
    post:
      operationId: createSkill
      parameters: [{ $ref: '#/components/parameters/XCompanyId' }]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/SkillInput' } } } }
      responses: { "201": { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Skill' } } } } }
  /skills/{id}:
    parameters:
      - { $ref: '#/components/parameters/XCompanyId' }
      - { $ref: '#/components/parameters/IdParam' }
    patch:
      operationId: updateSkill
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/SkillInput' } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Skill' } } } } }
  /skills/{id}/publish:
    parameters:
      - { $ref: '#/components/parameters/XCompanyId' }
      - { $ref: '#/components/parameters/IdParam' }
    post:
      operationId: publishSkill
      summary: >-
        Publishes the skill's current latest version (immutable afterward, same rule as
        agent versions) — PRD §6.6a.
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Skill' } } } } }
...